+++ scripts/main.js
@@ -2,6 +2,7 @@ import { MODULE_ID } from './constants.js';
 import { registerSettings } from './settings.js';
 import { registerAfflictionHooks } from './hooks/registration.js';
 import { detectSystem } from './systemCompat.js';
+import './regions/register.js';
 
 const MIGRATION_VERSION = 1;